refactor: move storage to lib/infra/
Part of #7697 — restructure src/lib/ into layered architecture.
Move storage module to lib/infra/storage/ as it is pure infrastructure.
Update all import paths (16 external + 10 internal + 1 config).

Closes #7725
