+++ turbo/apps/web/app/api/storages/prepare/route.ts
@@ -24,7 +24,7 @@ import {
   downloadManifest,
   verifyS3FilesExist,
 } from "../../../../src/lib/s3/s3-client";
-import { computeContentHashFromHashes } from "../../../../src/lib/storage/content-hash";
+import { computeContentHashFromHashes } from "../../../../src/lib/infra/storage/content-hash";
 import { env } from "../../../../src/env";
 import { logger } from "../../../../src/lib/logger";
