+++ turbo/apps/web/src/lib/blob/blob-service.ts
@@ -10,7 +10,7 @@ import {
   deleteS3Objects,
   downloadBlob as downloadBlobFromS3,
 } from "../s3/s3-client";
-import { hashFileContent, type FileEntry } from "../storage/content-hash";
+import { hashFileContent, type FileEntry } from "../infra/storage/content-hash";
 import { env } from "../../env";
 import { logger } from "../logger";
 import pLimit from "p-limit";
