+++ turbo/apps/web/src/lib/compose/server-side-compose.ts
@@ -6,7 +6,7 @@ import {
   type SupportedFramework,
 } from "@vm0/core";
 import type { AgentComposeYaml } from "../../types/agent-compose";
-import { uploadInstructionsServerSide } from "../storage/instruction-upload";
+import { uploadInstructionsServerSide } from "../infra/storage/instruction-upload";
 import { computeComposeVersionId } from "../agent-compose/content-hash";
 import {
   agentComposes,
