+++ turbo/apps/web/src/lib/run/executors/types.ts
@@ -1,4 +1,4 @@
-import type { StorageManifest } from "../../storage/types";
+import type { StorageManifest } from "../../infra/storage/types";
 import type { ResumeSession } from "../types";
 import type { ArtifactSnapshot } from "../../checkpoint/types";
 import type { Firewalls } from "@vm0/core";
