+++ turbo/apps/web/src/lib/session-history/session-history-service.ts
@@ -5,7 +5,7 @@
  */
 
 import { uploadBlobs, downloadBlob } from "../blob/blob-service";
-import { hashFileContent } from "../storage/content-hash";
+import { hashFileContent } from "../infra/storage/content-hash";
 import { logger } from "../logger";
 
 const log = logger("session-history");
